+++ src/GameCoreEngine.Application/DTOs/Requests/SubmitGuessRequest.cs
@@ -2,7 +2,7 @@ using GameCoreEngine.Domain.Entities;
 
 namespace GameCoreEngine.Application.DTOs.Requests;
 
-public record SubmitGuessRequest(Guid MatchId, Guid PlayerId, int X, int Y)
+public record SubmitGuessRequest(Guid MatchId, Guid PlayerId, double X, double Y)
 {
     public Coordinate ToCoordinate() => new(X, Y);
 }
